Integrate a Shell Command into Leaf
Provide more complete integration between Leaf and a sub-shell.

Leaf shell is implemented as an external command.  Additionally, the
handling of each supported shell type is handled by a subscript that
is written to know the details of that shell.

Supported shells are Bash and Zsh.

Extra config variables:

LEAF_SHELL_PROMPT_PREFIX   -- This defaults to lsh and is always
                              appended to the prompt.

LEAF_SHELL_ZSH_PROMPT_SIDE -- If set to the value 'right' makes this
                              support script append the Leaf prompt
                              info on the righthand side.  To reset
                              it back to the left, simply unset the
                              variable.

Resolves: LETOOLS-347
Change-Id: Ib3488ab7344ef07ee794429b66ffce00860446b9