Performance issue to detect external commands
To get the description from an external commands, leaf was executing the
file with --description arguments.
This was causing some performance issues since many of our extensions
are written in python and the python vm is fast but ran X times, it
create a latency.

So now the description must be in the 2048 first bytes of the file,
after the motif LEAF_DESCRIPTION.

Change-Id: I331d36a015bc8a27eca35bfc23ff799a8ac7ef77
Resolves: LETOOLS-409