Change Leaf's process for tagging latest
Leaf's tagging process for latest is now based on an identifier and
their tags. So, packages will now be compared only to other packages
that share the same identifier and the same tags. This allows better
differentiation between image types, i.e. LXSWI and SWI9X07Y or
onlycap, signed and unsigned. As well as, allow each type to have a
latest version.

Resolves: LETOOLS-1145
Change-Id: I59864a904784a7a76947777836b2e0427e62ba42